Extract own module per contract mutated
