+++ hydra-node/hydra-node.cabal
@@ -228,6 +228,9 @@ test-suite tests
     Hydra.API.ServerSpec
     Hydra.APISpec
     Hydra.BehaviorSpec
+    Hydra.Chain.Direct.Contract.Close
+    Hydra.Chain.Direct.Contract.CollectCom
+    Hydra.Chain.Direct.Contract.FanOut
     Hydra.Chain.Direct.Contract.Mutation
     Hydra.Chain.Direct.ContractSpec
     Hydra.Chain.Direct.Fixture
